services/compliance/zoll/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
nats-py==2.6.0
prometheus-client==0.17.1
httpx==0.25.2
pytest==8.1.1
pytest-asyncio==0.23.5
apscheduler==3.10.4
